Fix access of size_t as int

+++ src/disk_mgr.c
@@ -434,7 +434,7 @@ void disk_manager_rescan(void)
 		return;
 	}
 
-	printf("Found %d devices\n", globbuf.gl_pathc);
+	printf("Found %d devices\n", (int)globbuf.gl_pathc);
 
 	int glob_idx;
